# elderly-care-backend/docs/swagger.yaml
definitions:
...
  vo.ResponseVO:
    properties:
      data: {}
      msg:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: 养老服务平台接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: elderly-care-backend
  version: "1.0"
paths:
...
  /chat/contactList:
    get:
      consumes:
      - application/json
      description: 最近的联系人
      parameters:
      - description: 聊天时间:格式，2025-11-09T13:49:06.732+08:00(用于上滑加载，避免一次性加载过多，默认返回最近的聊过天的几个联系人)
        in: query
        name: lastChatTime
        type: string
      - description: 返回数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/vo.ResponseVO'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 最近的联系人
      tags:
      - 聊天模块
  /chat/record:
    get:
      consumes:
      - application/json
      description: 聊天记录
      parameters:
      - description: 聊天对象用户ID(用于上滑加载，避免一次性加载过多，默认返回最近的几条聊天信息)
        in: query
        name: accountID
        required: true
        type: integer
      - description: 消息ID
        in: query
        name: messageID
        type: integer
      - description: 返回数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/vo.ResponseVO'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 聊天记录
      tags:
      - 聊天模块
  /evaluation/account:
    get:
      consumes:
      - application/json
      description: 获取账号评价
      parameters:
      - description: 账号ID
        in: query
        name: accountID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/vo.ResponseVO'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 获取账号评价
      tags:
      - 评价模块
    put:
      consumes:
      - application/json
      description: 任务完成后对任务进行评价
      parameters:
      - description: 评价的账号ID
        in: query
        name: accountID
        required: true
        type: integer
      - description: 得分（1-5分）
        in: query
        name: score
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/vo.ResponseVO'
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 评价账号
      tags:
      - 评价模块
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"